Tests tests tests
All tests fixed to pass with the new encoding parameter.  I suspect because the encoding of the py files are UTF-8 the tests in this case need to use UTF-8 to get the same data back and forth

Deleted the "tests/tmp" dir and now it uses a tempfile that just gets wiped when it's done.
