--- tests/tmp/PLACEHOLDER.txt
@@ -1,1 +1,0 @@
-empty placeholder to make sure the directory exists on checkout
